Moved Widget functions to Independent Widget classes / Completed UI for Edit Profile Screen
