+++ lib/design/components/onboarding/exports.dart
@@ -1,3 +1,5 @@
+export 'additional_info_container.dart';
+export 'enter_your_bio.dart';
 export 'glint_custom_icon_chip.dart';
 export 'im_looking_for_card.dart';
 export 'interest_and_vibe_selection.dart';
